+++ src/main/java/dev/isxander/yacl3/gui/image/impl/AnimatedDynamicTextureImage.java
@@ -3,6 +3,7 @@ package dev.isxander.yacl3.gui.image.impl;
 import com.mojang.blaze3d.Blaze3D;
 import com.mojang.blaze3d.platform.NativeImage;
 import com.twelvemonkeys.imageio.plugins.webp.WebPImageReaderSpi;
+import dev.isxander.yacl3.debug.DebugProperties;
 import dev.isxander.yacl3.gui.image.ImageRendererFactory;
 import dev.isxander.yacl3.gui.utils.GuiUtils;
 import net.minecraft.CrashReport;
@@ -62,15 +63,14 @@ public class AnimatedDynamicTextureImage extends DynamicTextureImage {
         graphics.pose().translate(x, y, 0);
         graphics.pose().scale(ratio, ratio, 1);
 
-        GuiUtils.doTextureFiltering();
-
         GuiUtils.blitGuiTex(
                 graphics,
                 uniqueLocation,
                 0, 0,
                 frameWidth * currentCol, frameHeight * currentRow,
                 frameWidth, frameHeight,
-                this.width, this.height
+                this.width, this.height,
+                DebugProperties.IMAGE_FILTERING
         );
         graphics.pose().popPose();